Please excuse the item mess, that is getting cleaned up.
Merchants are now mostly functional.

-This client strictly enforces slot, so the DB will need to be changed to make sure each merchantlist starts at 0 AND are in order.

-Coin will update after you've zoned when you've made a purchase. We are missing the moneyupdate opcode for now.

-Stackable items display weird for now.

-I have not tested selling, but it may work.

Formatted opcode finally found. Lua text, guildmasters, shopkeepers, all now speak to you. Text on normal clients is broken for now.

Spell illusions now work, with some minor bugs. Zoning fixes everything though of course.

Added some opcodes and other misc stuff, maybe things might start working!

+++ common/eq_packet_structs.h
@@ -3114,9 +3114,9 @@ struct TraderClick_Struct{
 };
 
 struct FormattedMessage_Struct{
-	uint32	unknown0;
-	uint32	string_id;
-	uint32	type;
+	uint16	unknown0;
+	uint16	string_id;
+	uint16	type;
 	char	message[0];
 };
 struct SimpleMessage_Struct{
